experiments/llamaindex/requirements.txt
# Core
llama-index==0.11.23
llama-index-llms-ollama==0.3.4
llama-index-embeddings-ollama==0.3.1

# Hybrid retrieval (BM25 half of the fusion)
llama-index-retrievers-bm25==0.4.0

# Postgres + pgvector store (same DB as the Java app, different table)
llama-index-vector-stores-postgres==0.2.6
psycopg2-binary==2.9.10
asyncpg==0.29.0

# PDF support for SimpleDirectoryReader
pypdf==5.1.0

python-dotenv==1.0.1
